Rename Core task `Package your operator` to `Create operator manifests`
The `Package your operator` doc primarily discusses creating the CSV, as
opposed to packaging the created manifests to create an operator bundle
(which is the next step). This PR renames the doc to include a more
appropriate heading.
It also fixes the link in the doc to the Concepts/dependency-resolution
doc.