hack/ci: Support passing additional run options to a container engine
Update the `hack/ci/lint-check.sh` bash script, which is responsible for
linting the general HTML output code, and expose the ability to set
additional run options to docker. This is nice in the case where you
have selinux bits present in the cloned/forked repository locally, and
that results in the inability to properly read from the mounted volume
that contains the website source.