.github/workflows: Update the lint workflow to open an issue when the lint check fails on master
Update the lint workflow and add a job that checks whether the lint job
failed when it was triggered on a master branch push event. When that
happens, create an issue to help improve visibility into broken links on
the OLM's documentation site. Update the event triggers and run this
workflow on a cron expression, and enable the workflow_dispatch trigger.

Signed-off-by: timflannagan <timflannagan@gmail.com>