Disables external link hash checks
HTML proofer produces false positives for
links which contain hashes handled by JavaScript.

This is very common for GitHub links where we link
to a specific piece of code: e.g. `#L164-L180`
for lines from 164 to 180 in a specific file.
Pages like these do not have actual
`<a href="#L164-L180">...</a>` elements and
HTML proofer considers it to be an issue while
these links are perfectly ok.

Signed-off-by: Mikalai Radchuk <mradchuk@redhat.com>